feat(search): directly search processing reference
