+++ src/env.d.ts
@@ -1,5 +1,5 @@
 declare module "*.yml" {
-    declare const content: {[key: string]: unknown}
+    const content: {[key: string]: unknown}
 
     export default content
 }
